+++ src/main/java/nextstep/step3/RoomEscapeWebApplication.java
@@ -3,7 +3,6 @@ package nextstep.step3;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-
 @SpringBootApplication
 public class RoomEscapeWebApplication {
     public static void main(String[] args) {
