+++ src/main/java/nextstep/step3/reservation/controller/themeReservationController.java
@@ -1,10 +1,10 @@
-package nextstep.step3.controller;
+package nextstep.step3.reservation.controller;
 
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
-import nextstep.step3.dto.ReservationDetail;
-import nextstep.step3.dto.ReservationDto;
-import nextstep.step3.service.ThemeReservationService;
+import nextstep.step3.reservation.dto.ReservationDetail;
+import nextstep.step3.reservation.dto.ReservationDto;
+import nextstep.step3.reservation.service.ThemeReservationService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
