+++ src/main/java/nextstep/step3/theme/dao/ThemeDao.java
@@ -1,6 +1,6 @@
-package nextstep.step3.dao;
+package nextstep.step3.theme.dao;
 
-import nextstep.step3.entity.Theme;
+import nextstep.step3.theme.entity.Theme;
 
 import java.sql.SQLException;
