+++ src/main/java/nextstep/step3/theme/dto/ThemeDto.java
@@ -1,4 +1,4 @@
-package nextstep.step3.dto;
+package nextstep.step3.theme.dto;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
